refactor: merge psutil based observers
